+++ app/src/main/java/com/woowa/banchan/domain/usecase/cart/inter/DeleteCartUseCase.kt
@@ -1,0 +1,3 @@
+package com.woowa.banchan.domain.usecase.cart.inter
+
+interface DeleteCartUseCase
