+++ app/src/main/java/com/woowa/banchan/domain/usecase/cart/inter/UpdateCartCheckStateUseCase.kt
@@ -1,0 +1,3 @@
+package com.woowa.banchan.domain.usecase.cart.inter
+
+interface UpdateCartCheckStateUseCase
