+++ app/src/main/java/com/woowa/banchan/domain/usecase/cart/inter/UpdateCartCountUseCase.kt
@@ -1,0 +1,3 @@
+package com.woowa.banchan.domain.usecase.cart.inter
+
+interface UpdateCartCountUseCase
