--- app/src/main/java/com/woowa/banchan/domain/usecase/cart/interface/GetTotalFoodUseCase.kt
@@ -1,3 +1,0 @@
-package com.woowa.banchan.domain.usecase.cart.`interface`
-
-interface GetTotalFoodUseCase
