+++ app/src/main/java/com/woowa/banchan/domain/usecase/food/inter/GetDetailFoodUseCase.kt
@@ -1,3 +1,3 @@
 package com.woowa.banchan.domain.usecase.food.inter
 
-interface UpdateFoodCheckStateUseCase
+interface GetDetailFoodUseCase
