--- app/src/main/java/com/woowa/banchan/domain/usecase/food/inter/UpdateFoodCountUseCase.kt
@@ -1,3 +1,0 @@
-package com.woowa.banchan.domain.usecase.food.inter
-
-interface UpdateFoodCountUseCase
