+++ app/src/main/java/com/woowa/banchan/domain/usecase/order/inter/GetTotalOrderUseCase.kt
@@ -1,0 +1,3 @@
+package com.woowa.banchan.domain.usecase.order.inter
+
+interface GetTotalOrderUseCase
