--- app/src/main/java/com/woowa/banchan/domain/usecase/order/interface/GetTotalOrderUseCase.kt
@@ -1,3 +1,0 @@
-package com.woowa.banchan.domain.usecase.order.`interface`
-
-interface GetTotalOrderUseCase
