+++ app/src/main/java/com/woowa/banchan/domain/usecase/recent/inter/GetRecentlyViewedFoodsUseCase.kt
@@ -1,0 +1,3 @@
+package com.woowa.banchan.domain.usecase.recent.inter
+
+interface GetRecentlyViewedFoodsUseCase
