--- app/src/main/java/com/woowa/banchan/domain/usecase/recent/interface/GetRecentlyViewedFoodsUseCase.kt
@@ -1,3 +1,0 @@
-package com.woowa.banchan.domain.usecase.recent.`interface`
-
-interface GetRecentlyViewedFoodsUseCase
